requirements.txt
alembic==1.9.2
amqp==5.1.1
anyio==3.6.2
attrs==22.2.0
Automat==22.10.0
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==39.0.0
cssselect==1.2.0
fastapi==0.89.1
filelock==3.9.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
kombu==5.2.4
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==0.4.3
packaging==23.0
parsel==1.7.0
pathspec==0.11.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
Protego==0.2.1
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
PyDispatcher==2.0.6
pyOpenSSL==23.0.0
python-dotenv==0.21.1
pytz==2022.7.1
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
Scrapy==2.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.1
starlette==0.22.0
tldextract==3.4.0
tomli==2.0.1
Twisted==22.10.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
vine==5.0.0
w3lib==2.1.1
wcwidth==0.2.6
zope.interface==5.5.2
